finished.

+++ Vector/src/ru/academits/Vector/LV/vector/Vector.java
@@ -86,7 +86,7 @@ public class Vector {
         return this;
     }
 
-    public Vector multiplyByScalar(int scalar) {
+    public Vector multiplyByScalar(double scalar) {
         for (int i = 0; i < components.length; i++) {
             components[i] *= scalar;
         }
